Adjust changes to the review
