+++ javalin/src/main/java/io/javalin/http/sse/SseHandler.kt
@@ -4,10 +4,7 @@ import io.javalin.core.util.Header
 import io.javalin.http.Context
 import io.javalin.http.Handler
 import io.javalin.http.addListener
-import java.io.Closeable
 import java.util.function.Consumer
-
-internal fun interface CloseSseFunction : Closeable
 
 class SseHandler @JvmOverloads constructor(
     private val timeout: Long = 0,
