+++ pom.xml
@@ -100,7 +100,6 @@
         <common.io.version>2.11.0</common.io.version>
         <gson.version>2.8.9</gson.version>
         <java.websocket.version>1.5.2</java.websocket.version>
-        <!-- <javax.servlet.api.version>3.1.0</javax.servlet.api.version> -->
         <junit.version>5.7.2</junit.version>
         <hamcrest.version>1.3</hamcrest.version>
         <okhttp.version>4.9.1</okhttp.version>
